be/requirements.txt
strands-agents==1.10.0
strands-agents-tools[agent-core-memory]==0.2.9
strands-agents-builder==0.1.10
python-dotenv==1.1.1
bedrock-agentcore==0.1.7
pydantic>=2.0.0
mcp>=1.0.0
fastapi>=0.118.0
uvicorn>=0.32.0
aws-opentelemetry-distro~=0.12.1
